requirements.txt
amqp==5.3.1
asgiref==3.8.1
astroid==3.3.7
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dill==0.3.9
Django==4.2.17
djangorestframework==3.15.2
flake8==7.1.1
idna==3.10
isort==5.13.2
kombu==5.4.2
mccabe==0.7.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
pycodestyle==2.12.1
pyflakes==3.2.0
pylint==3.3.2
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tomlkit==0.13.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
